Reviewer round 3: trapping-source verification, corrected ranks, significant Spearman
- Verified every trapping value against open full texts (Murai 2012/2014,
  Zandarashvili 2020, Xue 2022, Wang 2024); removed unverifiable values:
  rucaparib 0.8x -> 1.0x (tie), veliparib 0.02x -> 0.1x (FA assay),
  niraparib 65x -> rank-only (no quantitative fold-change in open sources)
- Correlation: S1 span rho -0.975 exact p=0.033 (now significant);
  AAI rho -0.82 p=0.133; LOO -0.95..-1.00; honest power statement
- 21 script: explicit trap ranks w/ tie correction, hollow niraparib marker,
  panel B xlim fix; figures regenerated
- build_seed_dataset: rank column float (ties)

+++ scripts/build_seed_dataset.py
@@ -88,7 +88,7 @@ def main() -> int:
             "molecular_formula": c["molecular_formula"],
             "isomeric_smiles": smi, "inchikey_pubchem": ikey_pc,
             "trapping_class": c["trapping_class"],
-            "trapping_rank": int(c["trapping_rank"]),
+            "trapping_rank": float(c["trapping_rank"]),
             "parp1_selective": c["parp1_selective"] == "TRUE",
             "first_disclosure_year": year, "time_split": time_split,
             "primary_pmid": c["primary_pmid"],
